Effectively remove R_DirtyImage.
The variable still exists, since it's possible that packages might refer to it,
but it's always set to one.

The only functional impact of this change is that when saving the workspace when
R exits, it will always save the global environment, even in the unlikely case
that it is unmodified.


git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/branches/kmillar/trunk@1440 549f2c98-2231-0410-b618-a09cd5e121c0