Run native code in a PlainContext, not a BailoutContext.
This is required as while Closure::execute handles Bailouts correctly, the
generated code doesn't handle Bailouts when calling other functions.

This could also be fixed by generating code to handle Bailouts or by generating
code to switch to a PlainContext for the duration of any calls to other
functions, but those approaches require more effort and generate more and
slower code for the common case that all the control flow gets inlined in the
generated code.