Eliminate use of the SchwarzCounter for initializing global objects.
Even with the counter to ensure the ordering of object creation and destruction,
global objects with non-trivial constructors and destructors are too problematic
in practice.  This eliminates them.

The downside to this is that some of the subsystems need explicit
initialization, most notably calls to:
  Rf_InitMemory
  Rf_InitNames
  Rf_InitGlobalEnv
Code written to work with CR already does explicit initialization, so this only
affects a few CXXR unit tests.

Note that by the design of their constructor, GCRoots as global objects are
safe.