Use conservative stack scanning to identify stack roots, instead of requiring them to be explicitly protected, or generate a runtime map of root locations.
With this change, many uses of PROTECT and GCStackRoot are unnecessary.
The exception to this is objects that manage memory.  In that case, the
object must be explicitly protected to prevent deletion in situations where
the object is no longer referenced but there are still pointers to the managed
memory.