Change the representation of String.
The header describing the string (class String) and the characters are now
stored in a contiguous block of memory.

This allows the stack scanning code to recognize a pointer to the contents of
a string as being an internal pointer to the String.  This completely eliminates
the need to explicitly mark stack roots pointing to String values.

It also reduces the number of memory allocations and deallocations performed
by most string handling code and improves cache behaviour a bit.