Change the representation of FixedVector.
The header describing the vector (class FixedVector) and the data are now
stored in a contiguous block of memory for all lengths of vectors, not just
scalars.

This allows the stack scanning code to recognize a pointer to the contents of
a vector as being an internal pointer to the FixedVector.  This completely
eliminates the need to explicitly mark stack roots pointing to FixedVectors.

It also reduces the number of memory allocations and deallocations performed
by most vector handling code and improves cache behaviour a bit.