Add benchmarking script (#71)
* Add benchmarking script

This adds a Python script (bench/runbench.py) for benchmarking Rho. The
script attempts to automate the building and benchmarking of a specific
version of Rho. The goal is to make it simpler to gather performance
metrics and reduce the risk for human error affecting benchmark results.

The benchmark script takes as argument a Git reference for the version
of Rho to benchmark, it also benchmarks CR as a baseline for performance
comparison.

* Add benchmark report script

Added an R script to generate basic benchmark reports.

* Add more plots in report

* Update benchmark script

Added more benchmarks, and made script accept multiple git references on
command line.

* Generate benchmark report PDF with tables

* Use rho-devel/benchmarks as benchmark repository