Compatibility fixes to build dplyr.
Newly supported function in the API:
  SET_TYPEOF
  SET_OBJECT (no-op)
  STRING_PTR
  VECTOR_PTR (always produces a runtime error)
  HASHTAB    (always produces a runtime error)
  R_PromiseExpr
  R_ClosureExpr
  PREXPR     (macro)
  BODY_EXPR  (macro)
  applyClosure (macro)
  countContexts (macro to unimplemented function)
  DATAPTR    (only when USE_RINTERNALS is defined)

- Make R_NilValue a SEXP unless compiling rho.
- Make SEXPTYPE a typedef for int unless compiling rho.
- Remove definition of operator==() in exported R_ext/Complex.h
- Move length() from the rho namespace into the global namespace.

Additionally added commented out function declarations in Rinternals.h
for functions that are in the GNU R version, but missing in rho.

With these changes, dplyr and its dependencies can be compiled successfully.
Also, data.table can be compiled with the exception of a single line of
the form:
   TYPEOF(v) = ...
which if changed to:
   SET_TYPEOF(v, ...)
would compile without problems.