Add BlockPool allocator for GCNode allocation (#106)
* Add BlockPool allocator for GCNode allocation

This replaces the use of BDWGC for allocating GCNodes. The BlockPool
allocator allocates small objects in an arena, this speeds up allocation
and pointer lookup for scalars.

* Organize imports

* Remove sweep worklist

* Add copyright headers

* Fix includes: include <functional>

* Remove unused declaration and add comment

* Add more comments

* Replace deleted_bucket global by macro

* Move BlockPool to rho namespace

* Sweep worklist

This adds back the sweep worklist because it may affect the Travis
builds.

* Fix review comments + arena bounds initialization

* Add BlockPool support for Address Sanitizer

The Address Sanitizer poisoning has ben re-added for the BlockPool
allocator.

Unfortunately the freed object quarantine does not work well with the
new allocator since the allocator would still iterate aver the
quarantined objects as if they were actually live objects, resulting in
use-after-poison errors.

* Add TODO for adding a quarantine

* Implement quarantine in BlockPool and move ASAN code

* Remove redundant worklist

Removed the worklist that was used to avoid concurrent live object
iteration and deletion in the sweep phase.

* Split out hashtable searching to separate function

Refactored the code dealing with the hashtable so that a separate
function is used to lookup a pointer in the table. This simplifies
the code using the search a bit.

* Refactored allocation table into separate class

Moved the hashtable functions into a separate class to improve
encapsulation and readability.

* Refactor the new allocator

Refactored the new allocator into three separate classes:
GCNodeAllocator (the main class), AllocatorSuperblock, and
AllocationTable.

Fixed reviewer comments.

* Remove redundant declarations from header

* Refactoring and fixing reviewer comments

* Replaced macro constants by static constexpr.
* Fixed erros in comments and added a few more comments.
* Moved several top-level functions into classes.
* Abstracted out some pointer manipulation code into the relevant
  classes.
* Moved current allocation iteration for superblocks into the Superblock
  class.
* Use a maximum load factor of 50% to determine when the allocation
  table is resized, instead of a maximum collision count.

* Made several Allocation members private

* Group allocation size by class & simplify freelist

Refactored the handling of superblock sizes so that the size of the
blocks and the whole superblock is determined by m_size_class. This
replaces m_block_size and maked it possible to unify block allocation
and freeing accross different size classes - this was handled by
different member functions previously.

Removed the superblock-specific freelists. There is only a single
freelist array now which is indexed by size class. This greatly
simplified the freelist and quarantine handling by removing
special cases to handle different allocation sizes.

Refactored GCNodeAllocator::allocate() so that it directly tries to
allocate from a freelist, instead of delegating to AllocatorSuperblock.
Allocation of large non-superblock allocations is now handled in
GCNodeAllocator.

Changed naming style for all allocator classes so that it better maches
the existing Rho style.

* Add GCNodeAllocator tests, fix reviewer comments

* Added unit tests for GCNodeAllocator
(tests/rho/GCNodeAllocaotrTests.cpp).

* Rewrote much of AllocationTable. Most of the insert functions now
take an Allocation as argument. Aslo simplified the table resizing code.

* AllocatorSuperblock now initializes the arena using posix_memalign().

* Fixed some comments that were out of phase with the code.

* Removed the redundant AllocatorSuperblock::s_size_class[] array.

* Added some asserts to check preconditions.

* Add const modifiers to applicable functions

* Added const modifiers.
* Added comment documenting the next_log2 algorithm origin.
* Use anonymous namespace for ALLOCATION_CHECK map & functions.

* Fix review comments

Removed redundant ifdefs.

Removed redundant static modifiers for declarations in the anonymous
namespace.

Made arena_superblock_start, arena_superblock_end, arena_superblock_next
file-local.

* Disable freelists for large size classes

Large allocations are now freed rather than placed in a freelist.